UI/Modals: rename ModalResponse/Content to DialogResponse/Content
