+++ components/ILIAS/UI/src/Component/Input/Container/Form/Form.php
@@ -25,11 +25,11 @@ use ILIAS\Refinery\Transformation;
 use Psr\Http\Message\ServerRequestInterface;
 use ILIAS\UI\Component\Input\Container\Container;
 use ILIAS\UI\Component\Input\Input;
-use ILIAS\UI\Component\Modal\ModalContent;
+use ILIAS\UI\Component\Modal\DialogContent;
 
 /**
  * This describes commonalities between all forms.
  */
-interface Form extends Container, ModalContent
+interface Form extends Container, DialogContent
 {
 }
