+++ components/ILIAS/UI/src/Component/Link/Link.php
@@ -25,12 +25,12 @@ use ILIAS\UI\Component\HasContentLanguage;
 use ILIAS\Data\LanguageTag;
 use ILIAS\UI\Component\HasHelpTopics;
 use ILIAS\UI\Component\JavaScriptBindable;
-use ILIAS\UI\Component\Modal\ModalContent;
+use ILIAS\UI\Component\Modal\DialogContent;
 
 /**
  * Link base interface.
  */
-interface Link extends Component, HasContentLanguage, HasHelpTopics, JavaScriptBindable, ModalContent
+interface Link extends Component, HasContentLanguage, HasHelpTopics, JavaScriptBindable, DialogContent
 {
     /**
      * Get the action url of a link
