+++ components/ILIAS/UI/src/Component/MessageBox/MessageBox.php
@@ -21,12 +21,12 @@ declare(strict_types=1);
 namespace ILIAS\UI\Component\MessageBox;
 
 use ILIAS\UI\Component\Component;
-use ILIAS\UI\Component\Modal\ModalContent;
+use ILIAS\UI\Component\Modal\DialogContent;
 
 /**
  * Interface Message Box
  */
-interface MessageBox extends Component, ModalContent
+interface MessageBox extends Component, DialogContent
 {
     // Types of Message Boxes:
     public const FAILURE = "failure";
