+++ components/ILIAS/UI/src/Component/Modal/Dialog.php
@@ -31,12 +31,6 @@ use ILIAS\Data\URI;
 interface Dialog extends Component, JavaScriptBindable
 {
     /**
-     * Get a Dialog like this with content from this URI;
-     * The (ajax) call to the URI MUST return a ModalResponse.
-     */
-    public function withAsyncUrl(URI $url): self;
-
-    /**
      * Get the signal to load and show this Dialog Modal
      */
     public function getShowSignal(?URI $uri = null): Signal;
