+++ components/ILIAS/UI/src/Component/Modal/DialogContent.php
