+++ components/ILIAS/UI/src/Implementation/Component/Modal/DialogResponse.php
