+++ components/ILIAS/UI/src/templates/default/Modal/tpl.dialogresponse.html
@@ -1,0 +1,14 @@
+<div class="il-modal-response">
+    <section data-section="il-modal-response__title">
+        {TITLE}
+    </section>
+    <section data-section="il-modal-response__contents">
+        {CONTENT}
+    </section>
+    <section data-section="il-modal-response__buttons">
+        {BUTTONS}
+    </template>
+    <template data-section="il-modal-response__command">
+        {COMMAND}
+    </template>
+</div>
